fix: userid Long 형 변경

+++ src/main/java/com/mycom/myapp/users/dao/UserDao.java
@@ -13,7 +13,7 @@ public interface UserDao {
 	
 	List<UserDto> listUser();
 	
-	UserDto detailUser(int userId);
+	UserDto detailUser(Long userId);
 	
 	UserDto detailUserByEmail(String email);
